+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 Fixed
 =====
 
+- Handle non-existent file in the text file editor
 - Fix issue with Select widget not scrolling beyond selection with scrollbar
 
 ----